Resolve #10 cleanup codebase
Cleans up the codebase and removes duplicate code, resolves #10
