+++ src/app/page.tsx
@@ -3,6 +3,6 @@ import TestPage from "@/app/clientPages/TestPage"
 
 export default async function Page () {
   return (
-    <TestPage />
+    <MainPage />
   )
 }
